Refactor timer management into separate module with improved functionality
